requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
alembic==1.16.4
    # via flask-migrate
blinker==1.9.0
    # via flask
brotli==1.1.0
    # via fonttools
cachetools==5.5.2
    # via google-auth
certifi==2025.8.3
    # via requests
cffi==1.17.1
    # via weasyprint
charset-normalizer==3.4.3
    # via requests
click==8.2.1
    # via
    #   flask
    #   rq
cssselect2==0.8.0
    # via weasyprint
deprecated==1.3.1
    # via limits
et-xmlfile==2.0.0
    # via openpyxl
flask==3.1.1
    # via
    #   -r requirements.in
    #   flask-limiter
    #   flask-login
    #   flask-migrate
    #   flask-sqlalchemy
    #   flask-wtf
flask-limiter==4.1.1
    # via -r requirements.in
flask-login==0.6.3
    # via -r requirements.in
flask-migrate==4.1.0
    # via -r requirements.in
flask-sqlalchemy==3.1.1
    # via
    #   -r requirements.in
    #   flask-migrate
flask-wtf==1.2.2
    # via -r requirements.in
fonttools[woff]==4.59.0
    # via weasyprint
google-api-core==2.25.1
    # via google-api-python-client
google-api-python-client==2.178.0
    # via -r requirements.in
google-auth==2.40.3
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-auth-oauthlib
google-auth-httplib2==0.2.0
    # via
    #   -r requirements.in
    #   google-api-python-client
google-auth-oauthlib==1.2.2
    # via -r requirements.in
googleapis-common-protos==1.70.0
    # via google-api-core
gunicorn==23.0.0
    # via -r requirements.in
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.10
    # via requests
itsdangerous==2.2.0
    # via
    #   flask
    #   flask-wtf
jinja2==3.1.6
    # via flask
limits==5.8.0
    # via flask-limiter
mako==1.3.10
    # via alembic
markdown==3.10.2
    # via -r requirements.in
markupsafe==3.0.2
    # via
    #   flask
    #   jinja2
    #   mako
    #   werkzeug
    #   wtforms
numpy==2.3.2
    # via pandas
oauthlib==3.3.1
    # via requests-oauthlib
openpyxl==3.1.5
    # via -r requirements.in
ordered-set==4.1.0
    # via flask-limiter
packaging==25.0
    # via
    #   gunicorn
    #   limits
    #   pytesseract
pandas==2.3.1
    # via -r requirements.in
pillow==11.3.0
    # via
    #   pytesseract
    #   weasyprint
proto-plus==1.26.1
    # via google-api-core
protobuf==6.31.1
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
psycopg2-binary==2.9.10
    # via -r requirements.in
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycparser==2.22
    # via cffi
pydyf==0.11.0
    # via weasyprint
pyparsing==3.2.3
    # via httplib2
pyphen==0.17.2
    # via weasyprint
pytesseract==0.3.13
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.1.1
    # via -r requirements.in
pytz==2025.2
    # via pandas
redis==6.4.0
    # via
    #   -r requirements.in
    #   rq
requests==2.32.4
    # via
    #   google-api-core
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
rq==2.4.1
    # via -r requirements.in
rsa==4.9.1
    # via google-auth
six==1.17.0
    # via python-dateutil
sqlalchemy==2.0.42
    # via
    #   alembic
    #   flask-sqlalchemy
tinycss2==1.4.0
    # via
    #   cssselect2
    #   weasyprint
tinyhtml5==2.0.0
    # via weasyprint
typing-extensions==4.14.1
    # via
    #   alembic
    #   flask-limiter
    #   limits
    #   sqlalchemy
tzdata==2025.2
    # via pandas
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.5.0
    # via requests
weasyprint==66.0
    # via -r requirements.in
webencodings==0.5.1
    # via
    #   cssselect2
    #   tinycss2
    #   tinyhtml5
werkzeug==3.1.3
    # via
    #   flask
    #   flask-login
wrapt==2.1.2
    # via deprecated
wtforms==3.2.1
    # via flask-wtf
zopfli==0.2.3.post1
    # via fonttools
